requirements.txt
pytest
requests
allure-pytest
faker
selenium